Relocate redo components
